edita readme com codeSmells existentes

+++ projeto/professional.dart
@@ -9,8 +9,8 @@ class Professional implements IUser {
       _researchGrantBegin; //Day, month and year the student was added to the research grant
   String?
       _researchGrantDeadline; //Day, month and year of the research grant deadline
-  int _paymentValue = 0;
-  int _wallet = 0;
+  double _paymentValue = 0;
+  double _wallet = 0;
 
   // ####defining getters and setters####
 
@@ -68,12 +68,12 @@ class Professional implements IUser {
   }
 
   @override
-  set setPaymentValue(int newPaymentValue) {
+  set setPaymentValue(double newPaymentValue) {
     this._paymentValue = newPaymentValue;
   }
 
   @override
-  set setWallet(int newWallet) {
+  set setWallet(double newWallet) {
     this._wallet = newWallet;
   }
